separate error and non error resources
